fixes test suite bugs. adds JEST functionality
