+++ package.json
@@ -20,6 +20,7 @@
   },
   "homepage": "https://github.com/xillow-talk-hrsf110/steve-component-agent-contact#readme",
   "dependencies": {
+    "axios": "^0.18.0",
     "body-parser": "^1.18.3",
     "eslint-config-airbnb": "^17.1.0",
     "express": "^4.16.4",
